Fixed sonar.

+++ src/main/java/g0601_0700/s0692_top_k_frequent_words/Solution.java
@@ -23,7 +23,7 @@ public class Solution {
         SortedSet<Map.Entry<String, Integer>> sortedset =
                 new TreeSet<>(
                         (e1, e2) -> {
-                            if (e1.getValue() != e2.getValue()) {
+                            if (e1.getValue().intValue() != e2.getValue().intValue()) {
                                 return e2.getValue() - e1.getValue();
                             } else {
                                 return e1.getKey().compareToIgnoreCase(e2.getKey());
